+++ apps/extension/vite.config.ts
@@ -1,10 +1,25 @@
 import { defineConfig } from "vite";
 import react from "@vitejs/plugin-react";
 import tsconfigPaths from "vite-tsconfig-paths";
-import path from "path";
+import path, { resolve } from "path";
+import * as dotenv from "dotenv";
+import { viteStaticCopy } from "vite-plugin-static-copy";
+
+dotenv.config({ path: resolve(__dirname, "../../.env.shared") });
 
 export default defineConfig({
-  plugins: [react(), tsconfigPaths()],
+  plugins: [
+    react(),
+    tsconfigPaths(),
+    viteStaticCopy({
+      targets: [
+        {
+          src: resolve(__dirname, "../../packages/shared/assets/favicon.png"),
+          dest: "",
+        },
+      ],
+    }),
+  ],
   build: {
     target: "esnext",
     outDir: "dist",