+++ apps/web/src/constants/data/reviews.ts
@@ -1,0 +1,32 @@
+import type { ReviewModel } from "@models";
+
+export const REVIEWS: ReviewModel[] = [
+  {
+    firstName: "Stoica",
+    lastName: "Lavinia",
+    rating: 1,
+    review:
+      "View both your public and private IP addresses as soon as you open the extension — no clicks needed.",
+  },
+  {
+    firstName: "Morohoschi",
+    lastName: "Daniel",
+    rating: 2,
+    review:
+      "View both your public and private IP addresses as soon as you open the extension — no clicks needed.",
+  },
+  {
+    firstName: "Apostoaie",
+    lastName: "Alexandru",
+    rating: 3,
+    review:
+      "View both your public and private IP addresses as soon as you open the extension — no clicks needed.",
+  },
+  {
+    firstName: "Morohsochi",
+    lastName: "Camelia",
+    rating: 4,
+    review:
+      "View both your public and private IP addresses as soon as you open the extension — no clicks needed.",
+  },
+];