+++ apps/web/src/main.tsx
@@ -1,10 +1,21 @@
-import { StrictMode } from 'react'
-import { createRoot } from 'react-dom/client'
-import './index.css'
-import App from './App.tsx'
+import { StrictMode } from "react";
+import { createRoot } from "react-dom/client";
+import { Stack, VStack } from "@chakra-ui/react";
+import { Description, Features, Reviews } from "@sections";
+import { StyleProvider } from "@providers";
+import { Footer } from "@components";
 
-createRoot(document.getElementById('root')!).render(
+createRoot(document.getElementById("root")!).render(
   <StrictMode>
-    <App />
-  </StrictMode>,
-)
+    <StyleProvider>
+      <Stack gap="10">
+        <VStack width="full" gap="4" justifyContent="center">
+          <Description />
+          <Features />
+          <Reviews />
+        </VStack>
+        <Footer />
+      </Stack>
+    </StyleProvider>
+  </StrictMode>
+);