+++ apps/web/src/models/Feature.model.ts
@@ -1,0 +1,7 @@
+import type { IconType } from "react-icons/lib";
+
+export interface FeatureModel {
+  title: string;
+  description: string;
+  icon: IconType;
+}