+++ apps/web/src/models/Review.model.ts
@@ -1,0 +1,6 @@
+export interface ReviewModel {
+  firstName: string;
+  lastName: string;
+  rating: number;
+  review: string;
+}