+++ apps/web/vite.config.ts
@@ -1,7 +1,24 @@
-import { defineConfig } from 'vite'
-import react from '@vitejs/plugin-react'
+import { defineConfig } from "vite";
+import tsconfigPaths from "vite-tsconfig-paths";
+import react from "@vitejs/plugin-react";
+import * as dotenv from "dotenv";
+import { resolve } from "path";
+import { viteStaticCopy } from "vite-plugin-static-copy";
+
+dotenv.config({ path: resolve(__dirname, "../../.env.shared") });
 
 // https://vite.dev/config/
 export default defineConfig({
-  plugins: [react()],
-})
+  plugins: [
+    react(),
+    tsconfigPaths(),
+    viteStaticCopy({
+      targets: [
+        {
+          src: resolve(__dirname, "../../packages/shared/assets/favicon.png"),
+          dest: "",
+        },
+      ],
+    }),
+  ],
+});