+++ packages/shared/components/index.tsx
@@ -1,0 +1,2 @@
+export * from "./BuyMeACoffeeButton";
+export * from "./GitHubButton";